requirements.txt
PyYAML
requests
tinydb
tweepy